T2: extract TripTimeline into pure TripTimelineView + thin container (one inline PATCH lifted to callback)
